[refactor] refactor schedule code
