+++ flowjob-broker-application/src/main/java/org/limbo/flowjob/broker/application/component/schedule/ScheduleStrategyHelper.java
@@ -22,7 +22,6 @@ import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.BooleanUtils;
-import org.limbo.flowjob.broker.application.component.SlotManager;
 import org.limbo.flowjob.broker.core.dispatch.TaskDispatcher;
 import org.limbo.flowjob.broker.core.domain.IDGenerator;
 import org.limbo.flowjob.broker.core.domain.IDType;
@@ -35,27 +34,21 @@ import org.limbo.flowjob.broker.core.domain.task.TaskFactory;
 import org.limbo.flowjob.broker.core.exceptions.JobException;
 import org.limbo.flowjob.broker.dao.converter.DomainConverter;
 import org.limbo.flowjob.broker.dao.entity.JobInstanceEntity;
-import org.limbo.flowjob.broker.dao.entity.PlanEntity;
 import org.limbo.flowjob.broker.dao.entity.PlanInfoEntity;
 import org.limbo.flowjob.broker.dao.entity.PlanInstanceEntity;
-import org.limbo.flowjob.broker.dao.entity.PlanSlotEntity;
 import org.limbo.flowjob.broker.dao.entity.TaskEntity;
 import org.limbo.flowjob.broker.dao.repositories.JobInstanceEntityRepo;
 import org.limbo.flowjob.broker.dao.repositories.PlanEntityRepo;
 import org.limbo.flowjob.broker.dao.repositories.PlanInfoEntityRepo;
 import org.limbo.flowjob.broker.dao.repositories.PlanInstanceEntityRepo;
-import org.limbo.flowjob.broker.dao.repositories.PlanSlotEntityRepo;
 import org.limbo.flowjob.broker.dao.repositories.TaskEntityRepo;
 import org.limbo.flowjob.common.constants.JobStatus;
 import org.limbo.flowjob.common.constants.MsgConstants;
-import org.limbo.flowjob.common.constants.PlanStatus;
 import org.limbo.flowjob.common.constants.PlanType;
 import org.limbo.flowjob.common.constants.ScheduleType;
 import org.limbo.flowjob.common.constants.TaskStatus;
 import org.limbo.flowjob.common.constants.TaskType;
-import org.limbo.flowjob.common.constants.TriggerType;
 import org.limbo.flowjob.common.exception.VerifyException;
-import org.limbo.flowjob.common.utils.Verifies;
 import org.limbo.flowjob.common.utils.attribute.Attributes;
 import org.limbo.flowjob.common.utils.dag.DAG;
 import org.limbo.flowjob.common.utils.dag.DAGNode;
@@ -65,14 +58,12 @@ import org.springframework.stereotype.Component;
 
 import javax.inject.Inject;
 import javax.transaction.Transactional;
-import java.text.MessageFormat;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
@@ -95,9 +86,6 @@ public class ScheduleStrategyHelper {
     private TaskFactory taskFactory;
 
     @Setter(onMethod_ = @Inject)
-    private IDGenerator idGenerator;
-
-    @Setter(onMethod_ = @Inject)
     private PlanInstanceEntityRepo planInstanceEntityRepo;
 
     @Setter(onMethod_ = @Inject)
@@ -113,75 +101,10 @@ public class ScheduleStrategyHelper {
     private TaskDispatcher taskDispatcher;
 
     @Setter(onMethod_ = @Inject)
-    private SlotManager slotManager;
-
-    @Setter(onMethod_ = @Inject)
-    private PlanSlotEntityRepo planSlotEntityRepo;
-
-    @Setter(onMethod_ = @Inject)
     private JobInstanceHelper jobInstanceHelper;
 
-    @Transactional
-    public PlanInstanceEntity lockAndSavePlanInstance(Plan plan, TriggerType triggerType, LocalDateTime triggerAt) {
-        String planId = plan.getPlanId();
-        String version = plan.getVersion();
-
-        PlanEntity planEntity = planEntityRepo.selectForUpdate(planId);
-        Verifies.notNull(planEntity, MsgConstants.CANT_FIND_PLAN + planId);
-        Verifies.verify(!planEntity.isDeleted(), "plan:" + planId + " is deleted!");
-        Verifies.verify(planEntity.isEnabled(), "plan:" + planId + " is not enabled!");
-        // 任务是由之前时间创建的 调度时候如果版本改变 可能会有调度时间的变化本次就无需执行
-        // 比如 5s 执行一次 分别在 5s 10s 15s 在11s的时候内存里下次执行为 15s 此时修改为 2s 执行一次 那么重新加载plan后应该为 12s 14s 所以15s这次可以跳过
-        Verifies.verify(Objects.equals(version, planEntity.getCurrentVersion()), MessageFormat.format("plan:{0} version {1} change to {2}", planId, version, planEntity.getCurrentVersion()));
-
-        PlanInfoEntity planInfoEntity = planInfoEntityRepo.findById(version).orElseThrow(VerifyException.supplier(MessageFormat.format("does not find {0} plan info by version {1}", planId, version)));
-
-        // 判断是否由当前节点执行
-        if (TriggerType.API != triggerType) {
-
-            Verifies.verify(planEntity.isEnabled(), "plan " + planId + " is not enabled");
-
-            List<Integer> slots = slotManager.slots();
-            Verifies.notEmpty(slots, "slots is empty");
-            PlanSlotEntity planSlotEntity = planSlotEntityRepo.findByPlanId(planId);
-            Verifies.notNull(planSlotEntity, "plan's slot is null id:" + planId);
-            Verifies.verify(slots.contains(planSlotEntity.getSlot()), MessageFormat.format("plan {0} is not in this broker", planId));
-
-            // 校验是否重复创建
-            PlanInstanceEntity planInstanceEntity = planInstanceEntityRepo.findLatelyTrigger(planId, planInfoEntity.getPlanInfoId(), planInfoEntity.getScheduleType(), triggerType.type);
-            ScheduleType scheduleType = ScheduleType.parse(planInfoEntity.getScheduleType());
-            switch (scheduleType) {
-                case FIXED_RATE:
-                case CRON:
-                    Verifies.verify(planInstanceEntity == null || !triggerAt.isEqual(planInstanceEntity.getTriggerAt()),
-                            MessageFormat.format("Duplicate create PlanInstance,triggerAt:{0} planId[{1}] Version[{2}] oldPlanInstance[{3}]",
-                                    triggerAt, planId, version, JacksonUtils.toJSONString(planInstanceEntity))
-                    );
-                    break;
-                case FIXED_DELAY:
-                    Verifies.verify(planInstanceEntity == null || (!triggerAt.isEqual(planInstanceEntity.getTriggerAt()) && PlanStatus.parse(planInstanceEntity.getStatus()).isCompleted()),
-                            MessageFormat.format("Please wait last PlanInstance[{0}] complete.Plan[{1}] Version[{2}]",
-                                    JacksonUtils.toJSONString(planInstanceEntity), planId, version)
-                    );
-                    break;
-                default:
-                    throw new VerifyException(MsgConstants.UNKNOWN + " scheduleType " + planInfoEntity.getScheduleType());
-            }
-        }
-
-
-        String planInstanceId = idGenerator.generateId(IDType.PLAN_INSTANCE);
-        PlanInstanceEntity planInstanceEntity = new PlanInstanceEntity();
-        planInstanceEntity.setPlanInstanceId(planInstanceId);
-        planInstanceEntity.setPlanId(planId);
-        planInstanceEntity.setPlanInfoId(version);
-        planInstanceEntity.setStatus(PlanStatus.SCHEDULING.status);
-        planInstanceEntity.setTriggerType(triggerType.type);
-        planInstanceEntity.setScheduleType(planInfoEntity.getScheduleType());
-        planInstanceEntity.setTriggerAt(triggerAt);
-        planInstanceEntityRepo.saveAndFlush(planInstanceEntity);
-        return planInstanceEntity;
-    }
+    @Setter(onMethod_ = @Inject)
+    private IDGenerator idGenerator;
 
     @Transactional
     public JobInstance lockAndSaveWorkflowJobInstance(Plan plan, String planInstanceId, WorkflowJobInfo jobInfo, LocalDateTime triggerAt) {
@@ -315,13 +238,17 @@ public class ScheduleStrategyHelper {
         // 是否需要重试
         long retry = jobInstanceEntityRepo.countByPlanInstanceIdAndJobId(planInstanceId, jobInfo.getId());
         if (jobInfo.getRetryOption().getRetry() > retry) {
-            jobInstanceHelper.retryReset(jobInstance, jobInfo.getRetryOption().getRetryInterval());
+            String newJobInstanceId = idGenerator.generateId(IDType.JOB_INSTANCE);
+            jobInstance.retryReset(newJobInstanceId, jobInfo.getRetryOption().getRetryInterval());
             saveAndScheduleJobInstances(Collections.singletonList(jobInstance), TimeUtils.currentLocalDateTime());
         } else {
             handlerPlanComplete(planInstanceId, false);
         }
     }
 
+    /**
+     * 处理 map job 执行成功
+     */
     private void handleMapJobSuccess(Task task, JobInstance jobInstance) {
         switch (task.getType()) {
             case SPLIT:
@@ -335,6 +262,9 @@ public class ScheduleStrategyHelper {
         }
     }
 
+    /**
+     * 处理 map/reduce job 执行成功
+     */
     private void handleMapReduceJobSuccess(Task task, JobInstance jobInstance) {
         switch (task.getType()) {
             case SPLIT:
@@ -515,6 +445,7 @@ public class ScheduleStrategyHelper {
         taskEntityRepo.saveAll(taskEntities);
         taskEntityRepo.flush();
 
+        // task保存后才进行下发
         ScheduleStrategyContext.waitScheduleTasks(tasks.stream()
                 .map(task -> domainConverter.toTaskScheduleTask(task, triggerAt))
                 .collect(Collectors.toList())
