+++ flowjob-test/pom.xml
@@ -26,7 +26,7 @@
         <version>${flowjob.deploy.version}</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>test</artifactId>
+    <artifactId>flowjob-test</artifactId>
 
     <properties>
         <maven.deploy.skip>true</maven.deploy.skip>
